soluzione punto 1

+++ model/state.py
@@ -5,8 +5,8 @@ class State:
     id: str
     Name: str
     Capital: str
-    Lat: int
-    Lng: int
+    Lat: float
+    Lng: float
     Area: int
     Population: int
     Neighbors: str
